+++ Clinica-MVC/src/main/java/db/backend/clinicamvc/service/impl/PacienteService.java
@@ -1,6 +1,7 @@
 package db.backend.clinicamvc.service.impl;
 
 import db.backend.clinicamvc.entity.Paciente;
+import db.backend.clinicamvc.exception.ResourceNotFoundException;
 import db.backend.clinicamvc.repository.IPacienteRepository;
 import db.backend.clinicamvc.service.IPacienteService;
 import org.springframework.stereotype.Service;
@@ -35,8 +36,14 @@ public class PacienteService implements IPacienteService {
     }
 
     @Override
-    public void eliminarPaciente(Integer id) {
-        pacienteRepository.deleteById(id);
+    public void eliminarPaciente(Integer id) throws ResourceNotFoundException {
+        Optional<Paciente> pacienteOptional = buscarPorId(id);
+        if(pacienteOptional.isPresent()){
+            pacienteRepository.deleteById(id);
+        } else {
+            throw new ResourceNotFoundException("{\"message\": \"paciente no encontrado\"}");
+        }
+
     }
 
     @Override
